// frontend/src/app/shared/interfaces/amazon-search-params.ts
export class AmazonSearchParams {
  category?: string;
  priceFrom?: number;
  priceTo?: number;
  fromBSR?: number;
  toBSR?: number;
  reviewsFrom?: number;
  reviewsTo?: number;
  ratingFrom?: number;
  ratingTo?: number;
  numberOfSellersFrom?: number;
  numberOfSellersTo?: number;
  source: string;
  keyWord: string;
  checkSource: boolean;
  pageNumber: number;
  username?: string;
}